Исправила idNum, main.js
