+++ js/idNum.js
@@ -1,5 +1,13 @@
 //Массив из 25 чисел
-const idNum = new Array(25).fill(null);
-for (let int = 0; int < idNum.length - 1; int++) {
-  idNum.push(int);
-}
+const idNumInitial = new Array(25).fill(null);
+
+const fillArray = function (emptyArray) {
+  for (let int = 0; int < idNumInitial.length - 1; int++) {
+    idNumInitial.push(int);
+  }
+  return emptyArray;
+};
+
+const idNum = fillArray(idNumInitial);
+
+export {idNum};
